Replace Betanet with Mainnet Fork configuration
