+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/statecomputer/forks/ForksModule.java
@@ -36,7 +36,8 @@ public final class ForksModule extends AbstractModule {
 	@Override
 	protected void configure() {
 		OptionalBinder.newOptionalBinder(binder(), new TypeLiteral<UnaryOperator<Set<ForkConfig>>>() { });
-		install(new BetanetForkTxnRulesModule());
+		install(new MainnetForkConfigsModule());
+		install(new MainnetForkRulesModule());
 	}
 
 	@Provides
