+++ radixdlt-core/radixdlt/src/main/java/org/radix/GenerateUniverses.java
@@ -68,7 +68,7 @@ import com.radixdlt.statecomputer.RadixEngineModule;
 import com.radixdlt.statecomputer.checkpoint.Genesis;
 import com.radixdlt.statecomputer.checkpoint.GenesisProvider;
 import com.radixdlt.statecomputer.checkpoint.RadixNativeTokenModule;
-import com.radixdlt.statecomputer.forks.BetanetForkConfigsModule;
+import com.radixdlt.statecomputer.forks.MainnetForkConfigsModule;
 import com.radixdlt.store.EngineStore;
 import com.radixdlt.store.InMemoryEngineStore;
 import com.radixdlt.sync.CommittedReader;
@@ -290,7 +290,6 @@ public final class GenerateUniverses {
 					bind(UniverseType.class).toInstance(universeType);
 					install(new CryptoModule());
 					install(new RadixNativeTokenModule());
-					install(new BetanetForkConfigsModule());
 					install(new ForksModule());
 					install(RadixEngineConfig.asModule(1, 100, 50));
 					install(new RadixEngineModule());
