+++ radixdlt-core/radixdlt/src/test/java/com/radixdlt/api/chaos/mempoolfiller/MempoolFillerTest.java
@@ -47,7 +47,6 @@ import com.radixdlt.qualifier.NumPeers;
 import com.radixdlt.statecomputer.RadixEngineConfig;
 import com.radixdlt.statecomputer.RadixEngineStateComputer;
 import com.radixdlt.statecomputer.checkpoint.MockedGenesisModule;
-import com.radixdlt.statecomputer.forks.BetanetForkConfigsModule;
 import com.radixdlt.statecomputer.forks.RadixEngineForksLatestOnlyModule;
 import com.radixdlt.store.DatabaseLocation;
 
@@ -71,7 +70,6 @@ public class MempoolFillerTest {
 		return Guice.createInjector(
 			new RadixEngineForksLatestOnlyModule(new RERulesConfig(false, 100)),
 			MempoolConfig.asModule(10, 10),
-			new BetanetForkConfigsModule(),
 			new ForksModule(),
 			RadixEngineConfig.asModule(1, 100, 50),
 			new SingleNodeAndPeersDeterministicNetworkModule(),
