+++ radixdlt-core/radixdlt/src/test/java/com/radixdlt/api/service/SubmissionServiceTest.java
@@ -78,7 +78,6 @@ import com.radixdlt.statecomputer.TxnsCommittedToLedger;
 import com.radixdlt.statecomputer.checkpoint.Genesis;
 import com.radixdlt.statecomputer.checkpoint.MockedGenesisModule;
 import com.radixdlt.statecomputer.checkpoint.RadixEngineCheckpointModule;
-import com.radixdlt.statecomputer.forks.BetanetForkConfigsModule;
 import com.radixdlt.statecomputer.forks.RadixEngineForksLatestOnlyModule;
 import com.radixdlt.store.EngineStore;
 import com.radixdlt.store.InMemoryEngineStore;
@@ -138,7 +137,6 @@ public class SubmissionServiceTest {
 
 			@Override
 			public void configure() {
-				install(new BetanetForkConfigsModule());
 				install(new RadixEngineForksLatestOnlyModule(new RERulesConfig(false, 10)));
 				install(new ForksModule());
 				install(RadixEngineConfig.asModule(1, 100, 50));
