+++ radixdlt-core/radixdlt/src/test/java/com/radixdlt/recovery/RecoveryTest.java
@@ -76,7 +76,6 @@ import com.radixdlt.statecomputer.RadixEngineConfig;
 import com.radixdlt.statecomputer.RadixEngineModule;
 import com.radixdlt.statecomputer.checkpoint.Genesis;
 import com.radixdlt.statecomputer.checkpoint.MockedGenesisModule;
-import com.radixdlt.statecomputer.forks.BetanetForkConfigsModule;
 import com.radixdlt.statecomputer.forks.RadixEngineForksLatestOnlyModule;
 import com.radixdlt.store.DatabaseEnvironment;
 import com.radixdlt.store.DatabaseLocation;
@@ -141,7 +140,6 @@ public class RecoveryTest {
 		Guice.createInjector(
 			new MockedGenesisModule(),
 			new CryptoModule(),
-			new BetanetForkConfigsModule(),
 			new ForksModule(),
 			new RadixEngineForksLatestOnlyModule(new RERulesConfig(false, 100L)),
 			new RadixEngineModule(),
@@ -179,7 +177,6 @@ public class RecoveryTest {
 
 		return Guice.createInjector(
 			new RadixEngineForksLatestOnlyModule(new RERulesConfig(false, epochCeilingView)),
-			new BetanetForkConfigsModule(),
 			new ForksModule(),
 			MempoolConfig.asModule(10, 10),
 			RadixEngineConfig.asModule(1, Integer.MAX_VALUE, 50),
